Added Answer box
Added a box that shows up when the response is submitted. TODO: if response is incorrect, tell the right answer.
The Question View no longer has a Side Bar. Instead of clicking "Home" to quit a lesson, you click a new button. The button then prompts you if you really want to quit.